lndservices: move wallet_anchor.go into the new package
